# internal/swagger/spec.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "title": "durpdeploy REST API.",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "paths": {
...
    "/admin/tokens": {
      "get": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "admin-tokens"
        ],
        "summary": "List all API tokens (admin only).",
        "operationId": "listAllTokens",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "UserID",
            "name": "user_id",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Limit",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Offset",
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "TokenListResponse is a list of API tokens.",
            "schema": {
              "$ref": "#/definitions/TokenListResponse"
            }
          },
          "401": {
            "description": "Unauthorized error response.",
            "schema": {
              "$ref": "#/definitions/UnauthorizedError"
            }
          },
          "403": {
            "description": "Forbidden error response.",
            "schema": {
              "$ref": "#/definitions/ForbiddenError"
            }
          },
          "500": {
            "description": "Server error response.",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/tokens/{id}": {
      "delete": {
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "admin-tokens"
        ],
        "summary": "Revoke any API token by ID (admin only).",
        "operationId": "revokeAnyToken",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "TokenID",
            "description": "The API token ID.",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "EmptyResponse is returned for no-content endpoints.",
            "schema": {
              "$ref": "#/definitions/EmptyResponse"
            }
          },
          "401": {
            "description": "Unauthorized error response.",
            "schema": {
              "$ref": "#/definitions/UnauthorizedError"
            }
          },
          "403": {
            "description": "Forbidden error response.",
            "schema": {
              "$ref": "#/definitions/ForbiddenError"
            }
          },
          "404": {
            "description": "Not found error response.",
            "schema": {
              "$ref": "#/definitions/NotFoundError"
            }
          },
          "500": {
            "description": "Server error response.",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/tokens": {
      "get": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "tokens"
        ],
        "summary": "List API tokens for the authenticated user.",
        "operationId": "listTokens",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "UserID",
            "name": "user_id",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Limit",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Offset",
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "TokenListResponse is a list of API tokens.",
            "schema": {
              "$ref": "#/definitions/TokenListResponse"
            }
          },
          "401": {
            "description": "Unauthorized error response.",
            "schema": {
              "$ref": "#/definitions/UnauthorizedError"
            }
          },
          "500": {
            "description": "Server error response.",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "tokens"
        ],
        "summary": "Create a new API token.",
        "operationId": "createToken",
        "parameters": [
          {
            "x-go-name": "Body",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateTokenRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "CreateTokenResponse returns the plaintext token once.",
            "schema": {
              "$ref": "#/definitions/CreateTokenResponse"
            }
          },
          "400": {
            "description": "Bad request error response.",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "401": {
            "description": "Unauthorized error response.",
            "schema": {
              "$ref": "#/definitions/UnauthorizedError"
            }
          },
          "422": {
            "description": "Validation error response.",
            "schema": {
              "$ref": "#/definitions/ValidationError"
            }
          },
          "500": {
            "description": "Server error response.",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/tokens/{id}": {
      "delete": {
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "tokens"
        ],
        "summary": "Revoke the authenticated user's API token.",
        "operationId": "revokeToken",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "TokenID",
            "description": "The API token ID.",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "EmptyResponse is returned for no-content endpoints.",
            "schema": {
              "$ref": "#/definitions/EmptyResponse"
            }
          },
          "401": {
            "description": "Unauthorized error response.",
            "schema": {
              "$ref": "#/definitions/UnauthorizedError"
            }
          },
          "404": {
            "description": "Not found error response.",
            "schema": {
              "$ref": "#/definitions/NotFoundError"
            }
          },
          "500": {
            "description": "Server error response.",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "BadRequestError": {
      "type": "object",
      "title": "Bad request error response.",
      "properties": {
        "error": {
          "type": "string",
          "x-go-name": "Error"
        }
      },
      "x-go-name": "swaggerBadRequestError",
      "x-go-package": "durpdeploy/internal/handler/api"
    },
...
    "CreateTokenRequest": {
      "type": "object",
      "title": "CreateTokenRequest is the body for CreateToken.",
      "properties": {
        "name": {
          "type": "string",
          "x-go-name": "Name"
        }
      },
      "x-go-name": "swaggerCreateTokenRequest",
      "x-go-package": "durpdeploy/internal/handler/api"
    },
    "CreateTokenResponse": {
      "type": "object",
      "title": "CreateTokenResponse returns the plaintext token once.",
      "properties": {
        "created_at": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "CreatedAt"
        },
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "prefix": {
          "type": "string",
          "x-go-name": "Prefix"
        }
      },
      "x-go-name": "swaggerCreateTokenResponse",
      "x-go-package": "durpdeploy/internal/handler/api"
    },
...
    "EmptyResponse": {
      "type": "object",
      "title": "EmptyResponse is returned for no-content endpoints.",
      "x-go-name": "swaggerEmptyResponse",
      "x-go-package": "durpdeploy/internal/handler/api"
    },
...
    "ForbiddenError": {
      "type": "object",
      "title": "Forbidden error response.",
      "properties": {
        "error": {
          "type": "string",
          "x-go-name": "Error"
        }
      },
      "x-go-name": "swaggerForbiddenError",
      "x-go-package": "durpdeploy/internal/handler/api"
    },
...
    "NotFoundError": {
      "type": "object",
      "title": "Not found error response.",
      "properties": {
        "error": {
          "type": "string",
          "x-go-name": "Error"
        }
      },
      "x-go-name": "swaggerNotFoundError",
      "x-go-package": "durpdeploy/internal/handler/api"
    },
...
    "ServerError": {
      "type": "object",
      "title": "Server error response.",
      "properties": {
        "error": {
          "type": "string",
          "x-go-name": "Error"
        }
      },
      "x-go-name": "swaggerServerError",
      "x-go-package": "durpdeploy/internal/handler/api"
    },
...
    "TokenListResponse": {
      "type": "array",
      "title": "TokenListResponse is a list of API tokens.",
      "items": {
        "$ref": "#/definitions/TokenResponse"
      },
      "x-go-name": "swaggerTokenListResponse",
      "x-go-package": "durpdeploy/internal/handler/api"
    },
    "TokenResponse": {
      "type": "object",
      "title": "TokenResponse is the token representation returned by list endpoints.",
      "properties": {
        "created_at": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "CreatedAt"
        },
        "email": {
          "type": "string",
          "x-go-name": "Email"
        },
        "expires_at": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ExpiresAt"
        },
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "last_used_at": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "LastUsedAt"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "prefix": {
          "type": "string",
          "x-go-name": "Prefix"
        },
        "revoked_at": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "RevokedAt"
        },
        "scope": {
          "type": "string",
          "x-go-name": "Scope"
        },
        "user_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "UserID"
        },
        "user_name": {
          "type": "string",
          "x-go-name": "UserName"
        }
      },
      "x-go-name": "swaggerTokenResponse",
      "x-go-package": "durpdeploy/internal/handler/api"
    },
    "UnauthorizedError": {
      "type": "object",
      "title": "Unauthorized error response.",
      "properties": {
        "error": {
          "type": "string",
          "x-go-name": "Error"
        }
      },
      "x-go-name": "swaggerUnauthorizedError",
      "x-go-package": "durpdeploy/internal/handler/api"
    },
...
    "ValidationError": {
      "type": "object",
      "title": "Validation error response.",
      "properties": {
        "error": {
          "type": "string",
          "x-go-name": "Error"
        }
      },
      "x-go-name": "swaggerValidationError",
      "x-go-package": "durpdeploy/internal/handler/api"
    },
...
  },
  "securityDefinitions": {
    "bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}